+++ x/metadata/handler_test.go
@@ -1,0 +1,138 @@
+package metadata_test
+
+import (
+	"fmt"
+	"testing"
+
+	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
+
+	"github.com/provenance-io/provenance/app"
+	simapp "github.com/provenance-io/provenance/app"
+
+	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
+	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/stretchr/testify/suite"
+
+	testnet "github.com/cosmos/cosmos-sdk/testutil/network"
+	"github.com/provenance-io/provenance/x/metadata"
+	"github.com/provenance-io/provenance/x/metadata/types"
+	"github.com/provenance-io/provenance/x/metadata/types/p8e"
+)
+
+type HandlerTestSuite struct {
+	suite.Suite
+	cfg     testnet.Config
+	testnet *testnet.Network
+
+	app *app.App
+	ctx sdk.Context
+
+	pubkey1   cryptotypes.PubKey
+	user1     string
+	user1Addr sdk.AccAddress
+
+	pubkey2   cryptotypes.PubKey
+	user2     string
+	user2Addr sdk.AccAddress
+	handler   sdk.Handler
+}
+
+func (s *HandlerTestSuite) SetupTest() {
+	app := simapp.Setup(false)
+	s.app = app
+	ctx := app.BaseApp.NewContext(false, tmproto.Header{})
+	s.ctx = ctx
+
+	s.pubkey1 = secp256k1.GenPrivKey().PubKey()
+	s.user1Addr = sdk.AccAddress(s.pubkey1.Address())
+	s.user1 = s.user1Addr.String()
+
+	s.pubkey2 = secp256k1.GenPrivKey().PubKey()
+	s.user2Addr = sdk.AccAddress(s.pubkey2.Address())
+	s.user2 = s.user2Addr.String()
+
+	s.app.AccountKeeper.SetAccount(s.ctx, s.app.AccountKeeper.NewAccountWithAddress(s.ctx, s.user1Addr))
+
+	handler := metadata.NewHandler(app.MetadataKeeper)
+	s.handler = handler
+}
+
+func TestHandlerTestSuite(t *testing.T) {
+	suite.Run(t, new(HandlerTestSuite))
+}
+
+func createContractSpec(inputSpecs []*p8e.DefinitionSpec, outputSpec p8e.OutputSpec, definitionSpec p8e.DefinitionSpec) p8e.ContractSpec {
+	return p8e.ContractSpec{ConsiderationSpecs: []*p8e.ConsiderationSpec{
+		{FuncName: "additionalParties",
+			InputSpecs:       inputSpecs,
+			OutputSpec:       &outputSpec,
+			ResponsibleParty: 1,
+		},
+	},
+		Definition:      &definitionSpec,
+		InputSpecs:      inputSpecs,
+		PartiesInvolved: []p8e.PartyType{p8e.PartyType_PARTY_TYPE_AFFILIATE},
+	}
+}
+
+func createDefinitionSpec(name string, classname string, reference p8e.ProvenanceReference, defType int) p8e.DefinitionSpec {
+	return p8e.DefinitionSpec{
+		Name: name,
+		ResourceLocation: &p8e.Location{Classname: classname,
+			Ref: &reference,
+		},
+		Type: 1,
+	}
+}
+
+func (s HandlerTestSuite) TestAddContractSpecMsg() {
+	validDefSpec := createDefinitionSpec("perform_input_checks", "io.provenance.loan.LoanProtos$PartiesList", p8e.ProvenanceReference{Hash: "Adv+huolGTKofYCR0dw5GHm/R7sUWOwF32XR8r8r9kDy4il5U/LApxOWYHb05jhK4+eY4YzRMRiWcxU3Lx0+Mw=="}, 1)
+	invalidDefSpec := createDefinitionSpec("perform_action", "", p8e.ProvenanceReference{Hash: "Adv+huolGTKofYCR0dw5GHm/R7sUWOwF32XR8r8r9kDy4il5U/LApxOWYHb05jhK4+eY4YzRMRiWcxU3Lx0+Mw=="}, 1)
+
+	cases := map[string]struct {
+		v39CSpec p8e.ContractSpec
+		signers  []string
+		wantErr  bool
+		errorMsg string
+	}{
+		"should successfully ADD contract spec in from v38 to v40": {
+			createContractSpec([]*p8e.DefinitionSpec{&validDefSpec}, p8e.OutputSpec{Spec: &validDefSpec}, validDefSpec),
+			[]string{s.user1},
+			false,
+			"",
+		},
+		"should successfully UPDATE contract spec in from v38 to v40": {
+			createContractSpec([]*p8e.DefinitionSpec{&validDefSpec}, p8e.OutputSpec{Spec: &validDefSpec}, validDefSpec),
+			[]string{s.user1},
+			false,
+			"",
+		},
+		"should fail to add due to invalid signers": {
+			createContractSpec([]*p8e.DefinitionSpec{&validDefSpec}, p8e.OutputSpec{Spec: &validDefSpec}, validDefSpec),
+			[]string{s.user2},
+			true,
+			fmt.Sprintf("missing signature from existing owner %s; required for update", s.user1),
+		},
+		"should fail on converting contract validate basic": {
+			createContractSpec([]*p8e.DefinitionSpec{&invalidDefSpec}, p8e.OutputSpec{Spec: &validDefSpec}, validDefSpec),
+			[]string{s.user1},
+			true,
+			"input specification type name cannot be empty",
+		},
+	}
+	for n, tc := range cases {
+		tc := tc
+
+		s.Run(n, func() {
+			_, err := s.handler(s.ctx, &types.MsgAddP8EContractSpecRequest{Contractspec: tc.v39CSpec, Signers: tc.signers})
+			if tc.wantErr {
+				s.Error(err)
+				s.Equal(tc.errorMsg, err.Error())
+			} else {
+				s.NoError(err)
+			}
+		})
+	}
+
+}