+++ x/metadata/keeper/msg_server_test.go
@@ -1,8 +1,9 @@
 package keeper_test
 
 import (
-	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 	"testing"
+
+	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 
 	"github.com/provenance-io/provenance/app"
 	simapp "github.com/provenance-io/provenance/app"
@@ -13,15 +14,20 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/suite"
 
+	testnet "github.com/cosmos/cosmos-sdk/testutil/network"
+	"github.com/provenance-io/provenance/x/metadata/keeper"
 	"github.com/provenance-io/provenance/x/metadata/types"
 )
 
 type MsgServerTestSuite struct {
 	suite.Suite
+	cfg     testnet.Config
+	testnet *testnet.Network
 
 	app         *app.App
 	ctx         sdk.Context
 	queryClient types.QueryClient
+	msgServer   types.MsgServer
 
 	pubkey1   cryptotypes.PubKey
 	user1     string
@@ -34,7 +40,9 @@ type MsgServerTestSuite struct {
 
 func (s *MsgServerTestSuite) SetupTest() {
 	app := simapp.Setup(false)
+	s.app = app
 	ctx := app.BaseApp.NewContext(false, tmproto.Header{})
+	s.ctx = ctx
 
 	s.pubkey1 = secp256k1.GenPrivKey().PubKey()
 	s.user1Addr = sdk.AccAddress(s.pubkey1.Address())
@@ -46,10 +54,17 @@ func (s *MsgServerTestSuite) SetupTest() {
 
 	s.app.AccountKeeper.SetAccount(s.ctx, s.app.AccountKeeper.NewAccountWithAddress(s.ctx, s.user1Addr))
 
-	queryHelper := baseapp.NewQueryServerTestHelper(ctx, app.InterfaceRegistry())
+	queryHelper := baseapp.NewQueryServerTestHelper(s.ctx, app.InterfaceRegistry())
 	types.RegisterQueryServer(queryHelper, app.MetadataKeeper)
 	queryClient := types.NewQueryClient(queryHelper)
+
 	s.queryClient = queryClient
+	msgServer := keeper.NewMsgServerImpl(app.MetadataKeeper)
+	s.msgServer = msgServer
+
+	// TODO: Add a msgServer client to properly pass context
+	// for now use the handler_tests.go to test msg_server
+
 }
 
 func TestMsgServerTestSuite(t *testing.T) {